// src/user/dto/create-user.dto/create-personal_details.dto.ts
export class CreatePersonalDetailsDto {
  Looking_For: string[];

  Refreral_Partners: string[];

  DOB: Date;

  spouse_name: string;

  Hobbies: string[];

  Anniversary: Date;
  name: string;
  email: string;
  aboutus?: string;
  services?: any[];
  servicearea?: any[];
  Top_Clients?: any[];
  success_story: string;
  address: string;
}